requirements.txt
flask
gunicorn
pandas
numpy
pdfplumber
python-docx
scikit-learn
werkzeug
